Move 'detect_processing_qc_warnings' from 'qc.processing' to 'bcl2fastq.reporting'.
