+++ auto_process_ngs/commands/publish_qc_cmd.py
@@ -21,7 +21,7 @@ from ..docwriter import Link
 from ..docwriter import Para
 from ..docwriter import WarningIcon
 from ..barcodes.analysis import detect_barcodes_warnings
-from ..qc.processing import detect_processing_qc_warnings
+from ..bcl2fastq.reporting import detect_processing_qc_warnings
 from ..qc.utils import verify_qc
 from ..qc.utils import report_qc
 import bcftbx.utils as bcf_utils
